langgraph-scanner/requirements.txt
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backports.tarfile==1.2.0
boto3==1.42.0
botocore==1.41.6
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
deepl==1.23.0
distro==1.9.0
docstring_parser==0.17.0
docutils==0.22.2
exceptiongroup==1.3.0
executing==2.2.1
feedparser==6.0.12
filetype==1.2.0
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.71.0
groq==0.33.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
id==1.5.0
idna==3.11
importlib_metadata==8.7.0
ipython==8.37.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
jeepney==0.9.0
jiter==0.11.1
jmespath==1.0.1
joblib==1.5.2
jpype1==1.6.0
json_repair==0.52.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keyring==25.6.0
kiwipiepy==0.21.0
kiwipiepy_model==0.21.0
konlpy==0.6.0
langchain==1.0.1
langchain-anthropic==1.0.1
langchain-aws==1.1.0
langchain-core==1.1.0
langchain-google-genai==3.0.0
langchain-groq==1.0.0
langchain-mcp-adapters==0.1.11
langchain-openai==1.0.1
langchain-teddynote==0.5.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.37
langfuse[langchain]>=3.10.5
lxml==6.0.2
markdown-it-py==4.0.0
matplotlib-inline==0.1.7
mcp==1.18.0
mdurl==0.1.2
mmh3==4.1.0
more-itertools==10.8.0
nh3==0.3.1
nltk==3.9.2
numpy==2.3.5
olefile==0.47
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdf2image==1.17.0
pexpect==4.9.0
pillow==12.0.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.11.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
readme_renderer==44.0
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.16.0
SecretStorage==3.5.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tavily-python==0.7.12
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
traitlets==5.14.3
twine==6.2.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.12
uvicorn==0.38.0
wcwidth==0.2.14
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
